+++ CHANGES.rst
@@ -4,7 +4,8 @@ CHANGES
 `Unreleased <https://github.com/fmigneault/aiu/tree/master>`_ (latest)
 ------------------------------------------------------------------------------------
 
-* Nothing yet.
+* Fix parsing ``list`` format when number of lines can both result into 3-fields and 2-fields variant.
+* Fix handling unspecified ``--rename-format``, ``--rename-title`` and ``--prefix-track``.
 
 `0.5.0 <https://github.com/fmigneault/aiu/tree/0.5.0>`_ (2020-12-05)
 ------------------------------------------------------------------------------------